feat: Added TI / SYSTEM escalation
feat: added exceptions lists for reg values, reg keys and files 
fix: fixed permission errors
=> exceptions lists have been filled with values recreated at each boot

+++ src/registry.py
@@ -1,5 +1,33 @@
 import winreg as reg
 
+# Note : the following values and keys are ignored because they are recreated by the system at each reboot
+values_exeptions= {
+"HKCU\SOFTWARE\Microsoft\Windows\CurrentVersion\Explorer\Taskband": ["FavoritesResolve", "Favorites"],
+"HKLM\SYSTEM\CurrentControlSet\Control\WMI\Autologger\ReadyBoot" : "Start",
+"HKEY_LOCAL_MACHINE\SOFTWARE\Microsoft\Windows\CurrentVersion\Authentication\LogonUI\SessionData": "AllowLockScreen",
+}
+
+keys_exceptions = {
+"HKCU\SOFTWARE\Microsoft\Edge",
+"HKCU\SOFTWARE\Classes\Local Settings\Software\Microsoft\Windows\Shell\BagMRU",
+"HKCU\SOFTWARE\Classes\Local Settings\Software\Microsoft\Windows\Shell\Bags",
+"HKLM\SOFTWARE\Microsoft\Windows NT\CurrentVersion\Image File Execution Options"
+}
+
+
+def delete_key_with_sub_keys(key):
+    try:
+        i = 0
+        while True:
+            subkey_name = reg.EnumKey(key, i)
+            subkey = reg.OpenKey(key, subkey_name, 0, reg.KEY_ALL_ACCESS)
+            delete_key_with_sub_keys(subkey)
+            reg.DeleteKey(key, subkey_name)
+            reg.CloseKey(subkey)
+            i += 1
+    except WindowsError:
+        pass
+
 
 def getRegistryValue(key, value_name, deletion_detection=False):
     try:
@@ -67,14 +95,14 @@ def openRegistryKey(path):
         print(f"Permission denied to open {path}")
         return None
     except FileNotFoundError:
-        print(f"Could not find key {path}")
+        #print(f"Could not find key {path}")
         return None
     return key
 
 
 def checkAndResetValue(path, value_name, original_value, datatype, skip_prompts):
-    # Skipping these registry values as these two define the pinned values in the taskbar
-    if path == "HKCU\SOFTWARE\Microsoft\Windows\CurrentVersion\Explorer\Taskband" and (value_name == "FavoritesResolve" or value_name == "Favorites"):
+    # Skipping values that define user customized values (e.g. taskbar shortcuts)
+    if values_exeptions.get(path) and value_name in values_exeptions.get(path):
         return
     key = openRegistryKey(path)
     if key:
@@ -83,18 +111,21 @@ def checkAndResetValue(path, value_name, original_value, datatype, skip_prompts)
             original_value = bytes.fromhex(original_value)
         if (str(value) != str(original_value) and not (str(original_value) == "" and str(value) == "None")) and\
                 (skip_prompts or input(f"The registery value {value_name} at {path} is set to {str(value)} instead of {str(original_value)}. Do you want to reset it? (y/n) ") == 'y'):
-            print(f"Resetting registery value {value_name} at {path} to {original_value}")
+            print(f" ==> Resetting registery value {value_name} at {path} to {original_value}")
             setRegistryValue(key, value_name, original_value, datatype)
 
         reg.CloseKey(key)
 
 
 def checkKeyExistsAndDelete(path, skip_prompts):
+    if path in keys_exceptions:
+        return
     key = openRegistryKey(path)
     if key and (skip_prompts or input(f"The registery key {path} exists but should have been removed. Do you want to delete it? (y/n) ") == 'y'):
         try:
+            delete_key_with_sub_keys(key)
             reg.DeleteKey(key, "")
-            print(f"Deleting registery key {path}")
+            print(f" ==> Deleting registery key {path}")
         except PermissionError:
             print(f"Cannot delete regitry key {path}: Permission Error")
     reg.CloseKey(key)
@@ -106,5 +137,5 @@ def checkValueExistsAndDelete(path, value_name, skip_prompts):
         value = getRegistryValue(key, value_name, True)
         if value is not None and (skip_prompts or input(f"The registery value {value_name} at {path} is set to {str(value)} but should have been removed. Do you want to delete it? (y/n) ") == 'y'):
             reg.DeleteValue(key, value_name)
-            print(f"Deleting registery value {value_name} at {path}")
+            print(f" ==> Deleting registery value {value_name} at {path}")
         reg.CloseKey(key)
